requirements.txt
mysql-connector-python==8.0.25
pandas==1.2.4
psycopg2-binary==2.9.1
psycopg2-pool==1.1
postgres==3.0.0
python-dotenv==0.15.0